disk cache of translated urls (gt apy restarts often); use xsl's
New options:
--url-cache-path: where to store translated urls + source texts
--url-xsls: where to find checkout of gt {freecorpus,boundcorpus}/orig

The xsl's are used with pdfconverter for known pdfs needing special
case rules (but we might want to stick with pdftotext because of bugs
with spaces from pdftohtml, see
http://giellatekno.uit.no/bugzilla/show_bug.cgi?id=2378#c24 – in which
case we can throw out the xsl code).

We cache to disk and to memory on translation/retranslation.

When an url is not in memory but retrieved from the disk cache, the
cached version is sent back immediately, but then the original text is
retranslated in the background (since an apy restart typically means
the language pair was updated, so the next request will have a newer
version). The next request to the same pair+url will use the version
from memory.

The cache is now only used on 304's (timestamp is stored in the
cache), we send an If-Modified-Since header instead of the old hack of
just emptying the cache every 2 hrs.