Hardcode 4096 as pipebuf since PIPE_BUF on macos is artificially low
See https://github.com/afborchert/pipebuf#some-results
Darwin claims 512, but allows 65536.

8fbdd39d1bd798c6da7980b56e8b6d59e9576bd6 runs an equivalent of that C
program on startup to check the real pipebuf, should use that once
python 3.2 support is nixed. However, even if macos allows 65536, it
prefers 16384 and will switch to 512 if resoures are low, meaning
deadlocks are possible even if measurements reported 65536 as the
max. What do other tools do?

(In any case, we should have an api endpoint that reports pipebuf for
the js, so js can send chunks that are <pipebuf).