Clarify misleading 'Serving at' info message
http_server.bind() that follows the logging statements specifies just
the port, leaving address and family to their default values.

For the address argument, the default is None, meaning listening on all
available interfaces.

For the family argument, the default is AF_UNSPEC,
meaning both IPv4 and IPv6 will be used.

Clarify the comment to not mislead the user into believing the software
only binds on localhost IPv4/IPv6 addresses, but rather all interfaces
and families